refactor: 컨트롤러 및 서비스 레이어로 DTO 패키지 분리
