+++ src/main/java/roomescape/reservation/controller/dto/request/ReservationSaveRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto.request;
+package roomescape.reservation.controller.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
@@ -7,15 +7,16 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Status;
 import roomescape.reservation.domain.Theme;
+import roomescape.reservation.service.dto.request.ReservationPaymentRequest;
 
-public record ReservationRequest(
+public record ReservationSaveRequest(
         @NotNull Long memberId,
         @NotNull LocalDate date,
         @NotNull Long themeId,
         @NotNull Long timeId
 ) {
-    public static ReservationRequest of(ReservationPaymentSaveRequest detail, Long memberId) {
-        return new ReservationRequest(
+    public static ReservationSaveRequest of(ReservationPaymentSaveRequest detail, Long memberId) {
+        return new ReservationSaveRequest(
                 memberId,
                 detail.date(),
                 detail.themeId(),
@@ -23,8 +24,8 @@ public record ReservationRequest(
         );
     }
 
-    public static ReservationRequest from(ReservationPaymentRequest request) {
-        return new ReservationRequest(
+    public static ReservationSaveRequest from(ReservationPaymentRequest request) {
+        return new ReservationSaveRequest(
                 request.memberId(),
                 request.date(),
                 request.themeId(),
