+++ src/main/java/roomescape/reservation/controller/dto/response/PaymentResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto.response;
+package roomescape.reservation.controller.dto.response;
 
 public record PaymentResponse(String code, String message) {
 }
