+++ src/main/java/roomescape/reservation/service/ReservationTimeService.java
@@ -5,10 +5,10 @@ import java.time.LocalTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.reservation.controller.dto.request.TimeSaveRequest;
+import roomescape.reservation.controller.dto.response.AvailableReservationTimeResponse;
+import roomescape.reservation.controller.dto.response.TimeResponse;
 import roomescape.reservation.domain.ReservationTime;
-import roomescape.reservation.dto.request.TimeSaveRequest;
-import roomescape.reservation.dto.response.AvailableReservationTimeResponse;
-import roomescape.reservation.dto.response.TimeResponse;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
