+++ src/main/java/roomescape/reservation/service/ThemeService.java
@@ -4,11 +4,11 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.reservation.controller.dto.request.ThemeSaveRequest;
+import roomescape.reservation.controller.dto.response.PopularThemeResponse;
+import roomescape.reservation.controller.dto.response.ThemeResponse;
 import roomescape.reservation.domain.Theme;
 import roomescape.reservation.domain.ThemeName;
-import roomescape.reservation.dto.request.ThemeSaveRequest;
-import roomescape.reservation.dto.response.PopularThemeResponse;
-import roomescape.reservation.dto.response.ThemeResponse;
 import roomescape.reservation.repository.ThemeRepository;
 
 @Service
