+++ src/test/java/roomescape/common/validator/DateRangeValidatorTest.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import roomescape.reservation.dto.request.ReservationSearchCondRequest;
+import roomescape.reservation.controller.dto.request.ReservationSearchCondRequest;
 
 class DateRangeValidatorTest {
