+++ src/test/java/roomescape/reservation/controller/ReservationTimeControllerTest.java
@@ -14,7 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import roomescape.common.config.IntegrationTest;
 import roomescape.common.util.CookieUtils;
-import roomescape.reservation.dto.request.TimeSaveRequest;
+import roomescape.reservation.controller.dto.request.TimeSaveRequest;
 
 class ReservationTimeControllerTest extends IntegrationTest {
