+++ src/test/java/roomescape/reservation/controller/WaitingReservationControllerTest.java
@@ -21,7 +21,7 @@ import roomescape.common.config.IntegrationTest;
 import roomescape.common.util.CookieUtils;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberName;
-import roomescape.reservation.dto.request.WaitingReservationSaveRequest;
+import roomescape.reservation.controller.dto.request.WaitingReservationSaveRequest;
 
 class WaitingReservationControllerTest extends IntegrationTest {
