+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -24,15 +24,15 @@ import roomescape.auth.dto.LoginMember;
 import roomescape.common.config.DatabaseCleaner;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
+import roomescape.reservation.controller.dto.response.MemberReservationResponse;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.ReservationWithRank;
 import roomescape.reservation.domain.Theme;
-import roomescape.reservation.dto.request.ReservationPaymentRequest;
-import roomescape.reservation.dto.response.MemberReservationResponse;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.repository.ThemeRepository;
+import roomescape.reservation.service.dto.request.ReservationPaymentRequest;
 
 @SpringBootTest(webEnvironment = WebEnvironment.NONE)
 @ActiveProfiles("test")
