+++ src/test/java/roomescape/reservation/service/ReservationTimeServiceTest.java
@@ -25,14 +25,14 @@ import roomescape.common.config.DatabaseCleaner;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberName;
 import roomescape.member.repository.MemberRepository;
+import roomescape.reservation.controller.dto.request.TimeSaveRequest;
+import roomescape.reservation.controller.dto.response.AvailableReservationTimeResponse;
 import roomescape.reservation.domain.Description;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationTime;
 import roomescape.reservation.domain.Status;
 import roomescape.reservation.domain.Theme;
 import roomescape.reservation.domain.ThemeName;
-import roomescape.reservation.dto.request.TimeSaveRequest;
-import roomescape.reservation.dto.response.AvailableReservationTimeResponse;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.repository.ThemeRepository;
